added operator != for wchar_t
git-svn-id: svn://svn.code.sf.net/p/cxxtools/code/trunk@780 15ae5fad-cc11-0410-8fac-bce609e504b0

+++ cxxtools/include/cxxtools/char.h
@@ -222,7 +222,7 @@ namespace cxxtools {
             //! @brief Returns $true$ if the a and b are the same character; $false$ otherwise.
             //! @return $true$ if the a and b are the same character; $false$ otherwise.
             friend bool operator!=(const Char& a, wchar_t b)
-            { return a.value() != b; }
+            { return a.value() != (unsigned int)b; }
 
             //! @brief Returns $true$ if the a and b are the same character; $false$ otherwise.
             //! @return $true$ if the a and b are the same character; $false$ otherwise.
